Modularizando todos os meus contextos: separando os arquivos que vão exportar o useReducer e as initialStates de cada um
