+++ src/context/favoritos/FavoritosProvider.jsx
@@ -2,47 +2,9 @@ import { FavoritosContext } from "./FavoritosContext";
 
 import { useReducer, useEffect } from "react";
 
-// Definindo o reducer para gerenciar as ações de favoritos
-const favoritosReducer = (state, action) => {
-  switch (action.type) {
-    case "ADICIONAR_REMOVER_FAVORITO": {
-      const produtoExistente = state.favoritos.find(
-        (favorito) => favorito.id === action.payload.id
-      );
+import { favoritosReducer} from "./favoritosReducer";
 
-      if (produtoExistente) {
-        return {
-          ...state,
-          favoritos: state.favoritos.filter(
-            (favorito) => favorito.id !== action.payload.id
-          ),
-        };
-      } else {
-        return {
-          ...state,
-          favoritos: [...state.favoritos, action.payload],
-        };
-      }
-    }
-
-    case "LIMPAR_FAVORITOS":
-      return { ...state, favoritos: [] };
-
-    default:
-      return state;
-  }
-};
-
-// Função para obter os favoritos do localStorage
-const getFavoritosFromLocalStorage = () => {
-  const favoritosSalvos = localStorage.getItem("favorito");
-  return favoritosSalvos ? JSON.parse(favoritosSalvos) : []; // Retorna os favoritos salvos ou um array vazio
-};
-
-// Estado inicial do contexto
-const initialState = {
-  favoritos: getFavoritosFromLocalStorage(), // Carrega os favoritos salvos no localStorage
-};
+import { initialState } from "./initialState";
 
 export const FavoritosProvider = ({ children }) => {
   // Usando useReducer para gerenciar o estado dos favoritos
@@ -75,7 +37,7 @@ export const FavoritosProvider = ({ children }) => {
       dispatch({ type: "LIMPAR_FAVORITOS" });
     }
   };
-  
+
   return (
     <FavoritosContext.Provider
       value={{
