+++ src/context/produtos/ProdutosProvider.jsx
@@ -3,36 +3,9 @@ import { ProdutosContext } from "./ProdutosContext";
 
 import todosProdutos from "../../mocks/todosProdutos.json";
 
-const produtosReducer = (state, action) => {
-  switch (action.type) {
-    case "SELECIONAR_CATEGORIA":
-      return { ...state, categoriaSelecionada: action.payload };
-    case "DEFINIR_TERMO_BUSCA":
-      return { ...state, termoBusca: action.payload };
-    case "DEFINIR_FILTRO_PRECO":
-      return { ...state, filtroPreco: action.payload };
-    case "ABRIR_MODAL":
-      return { ...state, produtoAmpliado: action.payload };
-    case "FECHAR_MODAL":
-      return { ...state, produtoAmpliado: null };
-    case "LIMPAR_FILTROS":
-      return {
-        ...state,
-        categoriaSelecionada: null,
-        termoBusca: "",
-        filtroPreco: null,
-      };
-    default:
-      return state;
-  }
-};
+import { produtosReducer} from "./produtosReducer";
 
-const initialState = {
-  categoriaSelecionada: null,
-  termoBusca: "",
-  filtroPreco: null,
-  produtoAmpliado: null,
-};
+import { initialState } from "./initialState";
 
 export const ProdutoProvider = ({ children }) => {
   const [state, dispatch] = useReducer(produtosReducer, initialState);
