+++ src/context/produtos/initialState.js
@@ -1,0 +1,6 @@
+export const initialState = {
+    categoriaSelecionada: null,
+    termoBusca: "",
+    filtroPreco: null,
+    produtoAmpliado: null,
+  };
